refactor out telebot service
