+++ Pipfile
@@ -14,6 +14,7 @@ pytest-cov = "*"
 [packages]
 python-dotenv = "*"
 pytelegrambotapi = "*"
+fastapi = "*"
 
 [requires]
 python_version = "3.7"
