--- leetcode_tele_bot/handlers/__init__.py
@@ -1,2 +1,0 @@
-import leetcode_tele_bot.handlers.leetcode as leetcode
-import leetcode_tele_bot.handlers.telebot as telebot
