--- test.py
@@ -1,6 +1,0 @@
-import logging
-
-log = logging.getLogger("application")
-log.addHandler(logging.StreamHandler())
-log.setLevel(logging.DEBUG)
-log.info("Starting application")
